Also move IResultVisitor to output package
IDisplay is now the preferred interface to be implemented by clients
