+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/commands/IReplCommand.java
@@ -1,8 +1,8 @@
 package org.metaborg.spoofax.shell.commands;
 
-import org.metaborg.spoofax.shell.client.IResultVisitor;
 import org.metaborg.spoofax.shell.invoker.ICommandInvoker;
 import org.metaborg.spoofax.shell.output.IResult;
+import org.metaborg.spoofax.shell.output.IResultVisitor;
 
 /**
  * Interface for REPL commands. Used together with {@link ICommandInvoker}, instances of
