+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/FailOrSuccessResult.java
@@ -2,7 +2,6 @@ package org.metaborg.spoofax.shell.output;
 
 import java.util.Objects;
 
-import org.metaborg.spoofax.shell.client.IResultVisitor;
 import org.metaborg.spoofax.shell.functions.FailableFunction;
 
 /**
