+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/FailResult.java
@@ -1,7 +1,5 @@
 package org.metaborg.spoofax.shell.output;
 
-import org.metaborg.spoofax.shell.client.IResultVisitor;
-
 /**
  * An {@link IResult} that represents a failure in execution, along with the cause of the failure.
  */
