+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/IResult.java
@@ -2,7 +2,6 @@ package org.metaborg.spoofax.shell.output;
 
 import java.util.function.Consumer;
 
-import org.metaborg.spoofax.shell.client.IResultVisitor;
 import org.metaborg.spoofax.shell.commands.IReplCommand;
 
 /**
