+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/ISpoofaxResult.java
@@ -7,7 +7,6 @@ import org.apache.commons.vfs2.FileObject;
 import org.metaborg.core.context.IContext;
 import org.metaborg.core.messages.IMessage;
 import org.metaborg.core.unit.IUnit;
-import org.metaborg.spoofax.shell.client.IResultVisitor;
 import org.spoofax.interpreter.terms.IStrategoTerm;
 
 /**
